% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.vectors.list.R
\docType{data}
\name{example.vectors.list}
\alias{example.vectors.list}
\title{Example Vectors List}
\format{A list of named vectors.}
\usage{
example.vectors.list
}
\description{
An example dataset containing several named vectors, which can be compared to one another for overlaps, unique elements, etc.
}
\keyword{datasets}
