% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.random.colors.R
\name{generate.random.colors}
\alias{generate.random.colors}
\title{Generate Random Colors}
\usage{
generate.random.colors(number_of_colors_to_get)
}
\arguments{
\item{number_of_colors_to_get}{The number of colors to generate.}
}
\value{
A vector of R color names.
}
\description{
An internal function to generate a given number of random colors.
}
\examples{
generate_random_colors(5)
}
