% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.random.colors.R
\name{generate_random_colors}
\alias{generate_random_colors}
\title{Title}
\usage{
generate_random_colors(number_of_colors_to_get)
}
\arguments{
\item{number_of_colors_to_get}{}
}
\value{

}
\description{
Title
}
