% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/summarize.two.way.comparisons.percentage.overlap.R
\name{summarize.two.way.comparisons.percentage.overlap}
\alias{summarize.two.way.comparisons.percentage.overlap}
\title{Summarize Percentage Overlap for Two-Way Comparisons between Vectors}
\usage{
summarize.two.way.comparisons.percentage.overlap(named_list_of_vectors,
  output_type = "table", melt_table = FALSE, network_graph_minimum = 0)
}
\arguments{
\item{named_list_of_vectors}{(As in \code{\link{compare.vectors}}.)}

\item{output_type}{Either \code{"table"}, \code{"matrix_plot"}, or \code{"network_graph"}. \code{"table"} will return a matrix showing percentage overlap between each pair of vectors. \code{"matrix_plot"} will plot this table, coloring it by the amount of overlap. \code{"network_graph"} will return a network graph image illustrating the overlap percentages between each pair of vectors.}

\item{melt_table}{A logical (TRUE/FALSE) indicator, when \code{output_type} is \code{"table"}, whether to print the output in \code{\link[reshape2]{melt}ed} form (using the \pkg{reshape2} package).}

\item{network_graph_minimum}{\code{minimum} argument from \code{\link[qgraph]{qgraph}}, for when \code{output_type} is \code{"network_graph"}.}
}
\value{
Either a matrix (if \code{output} is \code{"table"}), or an image (if \code{output} is \code{"matrix_plot"} or \code{"network_graph"}). If an image is printed, nothing is returned by the function; rather, the output is printed immediately.

If \code{output} is \code{"table"} and \code{melt_table} is \code{FALSE}, the output will be a matrix with \code{nrow} and \code{ncol} both equal to the number of vectors in \code{named_list_of_vectors}. This table shows the decimal percentage overlap (e.g., "0.20" = 20\%) between each combination of vectors. \emph{This table is intended to be read with column names first, in this form:} "[column title] overlaps with [row title] [cell value] percent."

If \code{output} is \code{"table"} and \code{melt_table} is \code{TRUE}, the output will be a \code{\link[reshape2]{melt}ed} data.frame with three columns: \code{Vector_Name}, \code{Overlaps_With}, and \code{Decimal_Percentage}.
}
\description{
Summarize Percentage Overlap for Two-Way Comparisons between Vectors
}
\examples{
summarize.two.way.comparisons.percentage.overlap(veccompare::example.vectors.list)
summarize.two.way.comparisons.percentage.overlap(
	veccompare::example.vectors.list,
	output_type = "table",
	melt_table = TRUE
)

summarize.two.way.comparisons.percentage.overlap(
	veccompare::example.vectors.list,
	output_type = "matrix_plot"
)

summarize.two.way.comparisons.percentage.overlap(
	veccompare::example.vectors.list,
	output_type = "network_graph"
)
}
