% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.print.with.and.R
\name{vector.print.with.and}
\alias{vector.print.with.and}
\title{Print a vector with commas and a final "and"}
\usage{
vector.print.with.and(vector_to_print,
  string_to_return_if_vector_is_empty = "")
}
\arguments{
\item{vector_to_print}{A vector of strings (or elements able to be coerced into strings) to print.}

\item{string_to_return_if_vector_is_empty}{If \code{vector_to_print} is empty, the string that should be returned (for example, "", "(None)", etc.)}
}
\description{
Print a vector with commas and a final "and"
}
\examples{

vector.print.with.and(c("One", "Two", "Three", "Four"))
vector.print.with.and(c("One", "Two"))
vector.print.with.and(c("One"))
vector.print.with.and(c(), string_to_return_if_vector_is_empty = "(None)") # Outputs "(None)"
vector.print.with.and(c(""), string_to_return_if_vector_is_empty = "(None)") # Outputs ""
}
